webhook_app.py
# webhook/webhook_app.py

import os
from flask import Flask, request, jsonify
from googleapiclient.discovery import build
from googleapiclient.http import MediaIoBaseDownload, MediaFileUpload
from google.oauth2.service_account import Credentials
import io
import openai

app = Flask(__name__)

# Google Drive authentication NEW
def authenticate_google_drive():
    creds = Credentials.from_service_account_file(os.getenv("GOOGLE_CREDENTIALS_JSON_PATH"))
    drive_service = build("drive", "v3", credentials=creds)
    return drive_service

# Define your routes and logic here
@app.route("/webhook", methods=["POST"])
def handle_webhook():
    data = request.json
    file_id = data.get("file_id")
    file_name = data.get("file_name")
    target_language = data.get("target_language")

    if not all([file_id, file_name, target_language]):
        return jsonify({"status": "error", "message": "Missing required fields"}), 400

    try:
        # Authenticate Google Drive and download the file
        drive_service = authenticate_google_drive()
        # Implement download, translate, upload logic here
        # ...

        return jsonify({"status": "success", "translated_file_id": "some_id"})
    except Exception as e:
        return jsonify({"status": "error", "message": str(e)}), 500

if __name__ == "__main__":
    port = int(os.environ.get("PORT", 8080))
    app.run(host="0.0.0.0", port=port)